Run Geneva uploader test server integration in CI
