+++ opentelemetry-exporter-geneva/geneva-test-server/Cargo.toml
@@ -18,11 +18,19 @@ tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = ["env-filter", "fmt"] }
 uuid = { version = "1.0", features = ["v4", "serde"] }
 
-[dev-dependencies]
-geneva-uploader = { path = "../geneva-uploader", features = ["mock_auth"] }
-opentelemetry-proto = { workspace = true, default-features = false, features = ["logs", "gen-tonic-messages"] }
-reqwest = { version = "0.12", default-features = false, features = ["json", "rustls-tls"] }
-tempfile = "3.5"
+[features]
+testing = ["dep:reqwest", "dep:tempfile"]
+default = []
+
+[dependencies.reqwest]
+version = "0.12"
+default-features = false
+features = ["json", "rustls-tls"]
+optional = true
+
+[dependencies.tempfile]
+version = "3.5"
+optional = true
 
 [lints]
 workspace = true
